requirements.txt
# Smart Gold Trading Dashboard - Requirements
# Core Framework
streamlit==1.29.0
streamlit-option-menu==0.3.6

# Data Fetching and Processing
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2

# Visualization
plotly==5.18.0

# Technical Indicators
ta==0.11.0

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Authentication (Firebase)
firebase-admin==6.3.0
streamlit-authenticator==0.2.3
python-dotenv==1.0.0

# Utilities
requests==2.31.0
datetime

